Name unused file to _*
