+++ alea/_likelihoods/ll_nt_from_config.py
@@ -8,7 +8,7 @@ import alea.utils as utils
 import h5py
 import numpy as np
 import scipy.stats as sps
-from alea.plotting import pdf_plotter
+from alea._plotting import pdf_plotter
 from alea.simulators import simulate_interpolated
 from blueice.inference import bestfit_scipy, one_parameter_interval
 from blueice.likelihood import LogLikelihoodSum
