+++ alea/models/__init__.py
@@ -1,0 +1,2 @@
+from . import gaussian_model
+from .gaussian_model import *
