+++ tests/_test_VariationConvenience.py
